backend/requirements.txt
fastapi
uvicorn
httpx
sqlite3
psycopg2-binary
tiktoken
